# Improve the performance of Blocks\AssetsController::get_block_asset_resource_hints()