--- a/plugins/woocommerce/changelog/improve-performance-of-assetcontroller-get_script_dependency_src_array
+++ b/plugins/woocommerce/changelog/improve-performance-of-assetcontroller-get_script_dependency_src_array
@@ -1,0 +1,4 @@
+Significance: patch
+Type: performance
+
+Improve performance of uncached Block Assets dependency loading.