### Changes proposed in this Pull Request:

Closes [WOOPLUG-5559](https://linear.app/a8c/issue/WOOPLUG-5559/blockassetscontrollerget-block-asset-resource-hints-performance-issues) / #61014.

Improves upon #51874 which added caching around the handling.

This improves upon the performance of the underlying ::get_script_dependency_src_array() method by keeping track of visited nodes to reduce recursion since many dependencies have shared dependencies that should already be accounted for.


### Screenshots or screen recordings:

| Before | After | % Change |
| ------ | ----- | ----- |
|  140.57ms      |  2.02ms   | -98.% |

Before:

<img width="1002" height="580" alt="Screenshot 2025-09-18 at 3 00 52 PM" src="https://github.com/user-attachments/assets/dd8486b6-2dbc-427c-9bfa-5c6b028ed75a" />


After

<img width="2002" height="1163" alt="Screenshot 2025-09-18 at 3 21 48 PM" src="https://github.com/user-attachments/assets/f451f428-f122-4d33-9c9a-e2a8bf30b098" />

<!-- Begin testing instructions -->

### How to test the changes in this Pull Request:

<!-- Include detailed instructions on how these changes can be tested. Review and follow the guide for how to write high-quality testing instructions. -->

Using the [WooCommerce Testing Instructions Guide](https://github.com/woocommerce/woocommerce/wiki/Writing-high-quality-testing-instructions), include your detailed testing instructions:

1. Add something to your cart to trigger the preloading of cart block dependencies. 
2. Clear website cache to make sure the rendered output is from the new handling (object caching is required)
3. View the page source.  You should see a bunch of assets for prefetch:
<img width="1587" height="978" alt="Screenshot 2025-09-18 at 3 29 27 PM" src="https://github.com/user-attachments/assets/79c82f89-1091-492f-9aa6-b02fb5010c75" />

4. Save the output for comparison.
5. Re-checkout trunk.
6. Clear the website cache again.
7. View the page source again and save the ouptut.
8. Diff the two outputs to verify that no assets changed in any way.

<!-- End testing instructions -->